+++ docs/changelog.md
@@ -24,8 +24,8 @@ Release notes for `quimb`.
 **Bug fixes:**
 
 - fix [`isometrize_qr`](quimb.tensor.decomp.isometrize_qr) for complex torch arrays ({issue}`346`).
-
 - fix [`right_canonicalize`](quimb.tensor.tensor_1d.TensorNetwork1DFlat.right_canonicalize) to return the right canonicalized tensor network ({issue}`347`)
+- ensure all belief propagation contraction methods correctly propagate the target tensor network's `.exponent`.
 
 (whats-new-1-12-1)=
 ## v1.12.1 (2026-01-12)